+++ StereoPlus2/Windows.hpp
@@ -9,7 +9,6 @@
 #include <set>
 #include <sstream>
 #include <string>
-#include <filesystem> // C++17 standard header file name
 #include "include/imgui/imgui_stdlib.h"
 #include "FileManager.hpp"
 #include "TemplateExtensions.hpp"
@@ -18,7 +17,6 @@
 #include "ImGuiExtensions.hpp"
 #include "include/stb/stb_image_write.h"
 
-namespace fs = std::filesystem;
 
 //class TemplateWindow : Window {
 //	const Log log = Log::For<TemplateWindow>();
@@ -1217,38 +1215,6 @@ public:
 		Save,
 	};
 private:
-	class Path {
-		fs::path path;
-		std::string pathBuffer;
-	public:
-		const fs::path& get() {
-			return path;
-		}
-		std::string& getBuffer() {
-			return pathBuffer;
-		}
-
-		void apply() {
-			apply(pathBuffer);
-		}
-		void apply(fs::path n) {
-			path = fs::absolute(n);
-			pathBuffer = path.u8string();
-		}
-
-		bool isSome() {
-			return !pathBuffer.empty();
-		}
-
-		std::string join(Path path) {
-			auto last = pathBuffer[pathBuffer.size() - 1];
-		
-			if (last == '/' || last == '\\')
-				return pathBuffer + path.getBuffer();
-
-			return pathBuffer + '/' + path.getBuffer();
-		}
-	};
 
 	const Log log = Log::For<FileWindow>();
 
@@ -1420,20 +1386,42 @@ public:
 			ImGui::TreePop();
 		}
 
-		if (auto v = Settings::Language().Get();
-			ImGui::TreeNode(LocaleProvider::GetC("step"))) {
+		if (ImGui::TreeNode(LocaleProvider::GetC("step:step"))) {
 
-			if (auto v = Settings::TransitionStep().Get();
-				ImGui::InputFloat(LocaleProvider::GetC(Settings::Name(&Settings::TransitionStep)), &v, 0.01, 0.1))
-				Settings::TransitionStep() = v;
+			if (auto v = Settings::TranslationStep().Get();
+				ImGui::InputFloat(LocaleProvider::GetC("step:" + Settings::Name(&Settings::TranslationStep)), &v, 0.01, 0.1))
+				Settings::TranslationStep() = v;
 			if (auto v = Settings::RotationStep().Get();
-				ImGui::InputFloat(LocaleProvider::GetC(Settings::Name(&Settings::RotationStep)), &v, 0.01, 0.1))
+				ImGui::InputFloat(LocaleProvider::GetC("step:" + Settings::Name(&Settings::RotationStep)), &v, 0.01, 0.1))
 				Settings::RotationStep() = v;
-			if (auto v = Settings::ScaleStep().Get();
-				ImGui::InputFloat(LocaleProvider::GetC(Settings::Name(&Settings::ScaleStep)), &v, 0.01, 0.1))
-				Settings::ScaleStep() = v;
+			if (auto v = Settings::ScalingStep().Get();
+				ImGui::InputFloat(LocaleProvider::GetC("step:" + Settings::Name(&Settings::ScalingStep)), &v, 0.01, 0.1))
+				Settings::ScalingStep() = v;
 			ImGui::TreePop();
 		}
+
+		if (ImGui::TreeNode(LocaleProvider::GetC("color"))) {
+
+			if (auto v = Settings::ColorLeft().Get();
+				ImGui::ColorEdit4(LocaleProvider::GetC(Settings::Name(&Settings::ColorLeft)), (float*)&v, ImGuiColorEditFlags_NoInputs))
+				Settings::ColorLeft() = v;
+			if (auto v = Settings::ColorRight().Get();
+				ImGui::ColorEdit4(LocaleProvider::GetC(Settings::Name(&Settings::ColorRight)), (float*)&v, ImGuiColorEditFlags_NoInputs))
+				Settings::ColorRight() = v;
+			if (auto v = Settings::DimmedColorLeft().Get();
+				ImGui::ColorEdit4(LocaleProvider::GetC(Settings::Name(&Settings::DimmedColorLeft)), (float*)&v, ImGuiColorEditFlags_NoInputs))
+				Settings::DimmedColorLeft() = v;
+			if (auto v = Settings::DimmedColorRight().Get();
+				ImGui::ColorEdit4(LocaleProvider::GetC(Settings::Name(&Settings::DimmedColorRight)), (float*)&v, ImGuiColorEditFlags_NoInputs))
+				Settings::DimmedColorRight() = v;
+
+			ImGui::TreePop();
+		}
+
+		if (auto v = Settings::LogFileName().Get();
+			ImGui::InputText(LocaleProvider::GetC(Settings::Name(&Settings::LogFileName)), &v))
+			Settings::LogFileName() = v;
+		ImGui::SameLine(); ImGui::Extensions::HelpMarker("Requires restart.\n");
 
 		ImGui::End();
 		return true;
@@ -1442,4 +1430,24 @@ public:
 		return true;
 	}
 
+};
+
+class LogWindow : Window {
+	const Log log = Log::For<LogWindow>();
+public:
+	std::string Logs;
+
+	virtual bool Init() {
+		Window::name = "logWindow";
+
+		return true;
+	}
+	virtual bool Design() {
+		//ImGui::Extensions::LogTextMultiline(&Logs);
+		ImGui::InputTextMultiline("", &Logs, glm::vec2(-FLT_MIN, ImGui::GetTextLineHeight() * 16), ImGuiInputTextFlags_ReadOnly);
+		return true;
+	}
+	virtual bool OnExit() {
+		return true;
+	}
 };